Adaptation for git (#2)
Implements some minor changes necessary for git. Additionally, the index type for the author is changed to "tokenized", so that one can search for the surname or the email address, for example. Furthermore the parent commits are now indexed, too. Due to this, the index version is incremented.

See scm-manager/scm-manager#2111